+++ contracts/utils/LabelOperator.sol
@@ -57,14 +57,18 @@ abstract contract LabelOperator {
     //  MAX_LABEL_LENGTH < label.length < MIN_LABEL_LENGTH
     //  require(label.length > MAX_LABEL_LENGTH || label.length < MIN_LABEL_LENGTH, "INVALUD_LENGTH");
 
-    uint a = 0 ;
+    uint256 a = 0;
     for (uint256 i; i < label.length; i++) {
       require(
-        (label[i] >= bytes1("a") && label[i] <= bytes1("z")) || (label[i] >= bytes1("0") && label[i] <= bytes1("9")) || label[i] == bytes1("-") || label[i] == bytes1("_")||label[i] == 0x00,
+        (label[i] >= bytes1("a") && label[i] <= bytes1("z")) ||
+          (label[i] >= bytes1("0") && label[i] <= bytes1("9")) ||
+          label[i] == bytes1("-") ||
+          label[i] == bytes1("_") ||
+          label[i] == 0x00,
         "INVALID_CHARACTER"
       );
-      if(label[i] != 0x00){
-        a +=1;
+      if (label[i] != 0x00) {
+        a += 1;
       }
     }
     require(a < MAX_LABEL_LENGTH && a > MIN_LABEL_LENGTH, "INVALUD_LENGTH");
